# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API for managing caviar products and orders",
        "title": "Caviar API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/api/v1/orders": {
            "get": {
                "description": "Retrieve a list of orders with filtering and pagination",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "orders"
                ],
                "summary": "List orders",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Filter by order status",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by customer phone",
                        "name": "customer_phone",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by delivery country",
                        "name": "country",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by creation date from (RFC3339 format)",
                        "name": "created_from",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by creation date to (RFC3339 format)",
                        "name": "created_to",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "Page number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 10,
                        "description": "Items per page",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/caviar_internal_dto.OrderListResponseDTO"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "post": {
                "description": "Create a new order with customer and delivery information",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "orders"
                ],
                "summary": "Create a new order",
                "parameters": [
                    {
                        "description": "Order creation data",
                        "name": "order",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/caviar_internal_dto.OrderCreateDTO"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/caviar_internal_dto.OrderResponseDTO"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/api/v1/orders/number/{orderNumber}": {
            "get": {
                "description": "Retrieve an order by its order number",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "orders"
                ],
                "summary": "Get order by order number",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Order Number",
                        "name": "orderNumber",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/caviar_internal_dto.OrderResponseDTO"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
        "/api/v1/orders/statistics": {
            "get": {
                "description": "Get order statistics including counts by status and country",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "orders"
                ],
                "summary": "Get order statistics",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/api/v1/orders/{id}": {
            "get": {
                "description": "Retrieve an order by its ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "orders"
                ],
                "summary": "Get order by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Order ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/caviar_internal_dto.OrderResponseDTO"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete an order by ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "orders"
                ],
                "summary": "Delete order",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Order ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/api/v1/orders/{id}/status": {
            "put": {
                "description": "Update the status of an existing order",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "orders"
                ],
                "summary": "Update order status",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Order ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Status update data",
                        "name": "status",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/caviar_internal_dto.OrderStatusUpdateDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "caviar_internal_dto.CaviarDetailsDTO": {
            "type": "object",
            "properties": {
                "color": {
                    "type": "string"
                },
                "fishAge": {
                    "type": "string"
                },
                "grainSize": {
                    "type": "string"
                },
                "shelfLife": {
                    "$ref": "#/definitions/caviar_internal_dto.ShelfLifeDTO"
                },
                "taste": {
                    "type": "string"
                },
                "texture": {
                    "type": "string"
                }
            }
        },
        "caviar_internal_dto.CustomerInfoDTO": {
            "type": "object",
            "required": [
                "phone"
            ],
            "properties": {
                "email": {
                    "type": "string"
                },
                "firstName": {
                    "type": "string"
                },
                "fullName": {
                    "type": "string"
                },
                "lastName": {
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                }
            }
        },
        "caviar_internal_dto.DeliveryInfoDTO": {
            "type": "object",
            "required": [
                "city",
                "country",
                "type"
            ],
            "properties": {
                "address": {
                    "type": "string"
                },
                "city": {
                    "type": "string"
                },
                "country": {
                    "type": "string"
                },
                "instructions": {
                    "type": "string"
                },
                "postOffice": {
                    "type": "string"
                },
                "type": {
                    "type": "string",
                    "enum": [
                        "post_office",
                        "courier",
                        "address"
                    ]
                }
            }
        },
        "caviar_internal_dto.MoneyDTO": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "integer"
                },
                "currency": {
                    "type": "string"
                }
            }
        },
        "caviar_internal_dto.OrderCreateDTO": {
            "type": "object",
            "required": [
                "customerInfo",
                "deliveryInfo",
                "items"
            ],
            "properties": {
                "customerInfo": {
                    "$ref": "#/definitions/caviar_internal_dto.CustomerInfoDTO"
                },
                "deliveryInfo": {
                    "$ref": "#/definitions/caviar_internal_dto.DeliveryInfoDTO"
                },
                "items": {
                    "type": "array",
                    "minItems": 1,
                    "items": {
                        "$ref": "#/definitions/caviar_internal_dto.OrderItemDTO"
                    }
                },
                "notes": {
                    "type": "string"
                }
            }
        },
        "caviar_internal_dto.OrderItemDTO": {
            "type": "object",
            "required": [
                "productId",
                "quantity",
                "unitPrice",
                "variantId"
            ],
            "properties": {
                "productId": {
                    "type": "string"
                },
                "quantity": {
                    "type": "integer",
                    "minimum": 1
                },
                "unitPrice": {
                    "$ref": "#/definitions/caviar_internal_dto.MoneyDTO"
                },
                "variantId": {
                    "type": "string"
                }
            }
        },
        "caviar_internal_dto.OrderItemResponseDTO": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "product": {
                    "$ref": "#/definitions/caviar_internal_dto.ProductResponseDTO"
                },
                "productId": {
                    "type": "string"
                },
                "quantity": {
                    "type": "integer"
                },
                "totalPrice": {
                    "$ref": "#/definitions/caviar_internal_dto.MoneyDTO"
                },
                "unitPrice": {
                    "$ref": "#/definitions/caviar_internal_dto.MoneyDTO"
                },
                "variantId": {
                    "type": "string"
                }
            }
        },
        "caviar_internal_dto.OrderListResponseDTO": {
            "type": "object",
            "properties": {
                "limit": {
                    "type": "integer"
                },
                "orders": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/caviar_internal_dto.OrderResponseDTO"
                    }
                },
                "page": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                }
            }
        },
        "caviar_internal_dto.OrderResponseDTO": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "customerInfo": {
                    "$ref": "#/definitions/caviar_internal_dto.CustomerInfoDTO"
                },
                "deliveryInfo": {
                    "$ref": "#/definitions/caviar_internal_dto.DeliveryInfoDTO"
                },
                "id": {
                    "type": "string"
                },
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/caviar_internal_dto.OrderItemResponseDTO"
                    }
                },
                "notes": {
                    "type": "string"
                },
                "orderNumber": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "totalAmount": {
                    "$ref": "#/definitions/caviar_internal_dto.MoneyDTO"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "caviar_internal_dto.OrderStatusUpdateDTO": {
            "type": "object",
            "required": [
                "status"
            ],
            "properties": {
                "status": {
                    "type": "string",
                    "enum": [
                        "pending",
                        "confirmed",
                        "processing",
                        "shipped",
                        "delivered",
                        "cancelled"
                    ]
                }
            }
        },
...
        "caviar_internal_dto.ProductResponseDTO": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "details": {
                    "$ref": "#/definitions/caviar_internal_dto.CaviarDetailsDTO"
                },
                "id": {
                    "type": "string"
                },
                "is_active": {
                    "type": "boolean"
                },
                "name": {
                    "type": "string"
                },
                "slug": {
                    "type": "string"
                },
                "subtitle": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "variants": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/caviar_internal_dto.VariantResponseDTO"
                    }
                }
            }
        },
...
        "caviar_internal_dto.ShelfLifeDTO": {
            "type": "object",
            "properties": {
                "duration": {
                    "type": "string"
                },
                "tempRange": {
                    "$ref": "#/definitions/caviar_internal_dto.TemperatureRangeDTO"
                }
            }
        },
        "caviar_internal_dto.TemperatureRangeDTO": {
            "type": "object",
            "properties": {
                "maxC": {
                    "type": "number"
                },
                "minC": {
                    "type": "number"
                }
            }
        },
...
        "caviar_internal_dto.VariantResponseDTO": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "mass": {
                    "type": "integer"
                },
                "prices": {
                    "type": "object",
                    "additionalProperties": {
                        "$ref": "#/definitions/caviar_internal_dto.MoneyDTO"
                    }
                },
                "stock": {
                    "type": "integer"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}